requirements/test.txt
# Test dependencies go here.
-r base.txt
coverage==3.6
django-discover-runner==0.4
selenium
